mb/google/poppy/variants/nami: change type of board_sku_id() to uint32_t
Tools/scripts, like mosys/arc-setup, use int (4 bytes) to
read the sku id. In order to support "-1", we need to use
uint32_t (4 bytes) instead of using uint16_t (2 bytes) data type.
Otherwise, tools/scripts will read 65535 instead of -1.
Another reason to change this is that sku_id can be
supported by ec up to 4 bytes.

BUG=b:73792190
TEST=mosys output "Platform not supported" for -1 sku id
     arc-setup read -1 sku id

Reviewed-on: https://review.coreboot.org/25252
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>

+++ src/mainboard/google/poppy/variants/baseboard/include/baseboard/variants.h
@@ -53,7 +53,7 @@ struct memory_params {
 void variant_memory_params(struct memory_params *p);
 int variant_memory_sku(void);
 void variant_devtree_update(void);
-uint16_t variant_board_sku(void);
+uint32_t variant_board_sku(void);
 
 struct nhlt;
 void variant_nhlt_init(struct nhlt *nhlt);
